requirements.txt

fastapi>=0.111
uvicorn[standard]>=0.23
pydantic>=2.4
sqlmodel>=0.0.14
numpy>=1.24
scikit-learn>=1.2
scikit-image>=0.21
pillow>=10.0
umap-learn>=0.5
python-multipart>=0.0.6
typing_extensions>=4.9
joblib>=1.3
tensorflow-cpu>=2.15
astropy>=6.0
